Remove Base64 encoding from PUT / GET attachments. refs #10
Base64 encoding is only required when sending attachments inline. This
commit removes any base64 encoding / decoding around attachments until
there is explicit support for inline attachments. This ensures that
attachments are handled correctly (we will not change the bytes that the
consumer adds to a request) but also means we are no longer Base64
decoding attachments when they are fetched (because this is not required).

** note that this represents a breaking change for any existing attachments which
have been stored using MyCouch and incorrectly base64 encoded **